Former-commit-id: 8441533913a0bec068b4d60550c016e34512aec0 [formerly 793d9a8a50881ef8444fa3d8c0928d641ca017c6]
Former-commit-id: 557d25711b175b1994f82c1182b427aff20568a1

+++ src/vm/vm8051.cpp
@@ -943,7 +943,7 @@ static RET_DEFINE ExModifyDataDBVavleFunc(unsigned char * ipara,void * pVmEvn)
 		}
 	}
 
-	step =(int64_t)((*retdata.at(1)).size()- vTemp.size() -1);
+	step =(((int64_t)(*retdata.at(1)).size())- (int64_t)(vTemp.size()) -1);
 	auto tem =  make_shared<std::vector< vector<unsigned char> > >();
     CDataStream tep(SER_DISK, CLIENT_VERSION);
     tep << flag;
